requirements.txt
# S&P 500 Regime Detection AI Agent
# Python 3.12

# Data
yfinance>=0.2.36
pandas>=2.2
numpy>=1.26
pyarrow>=15.0

# Machine Learning
hmmlearn>=0.3.2
scikit-learn>=1.4
joblib>=1.3

# Visualization
matplotlib>=3.8
seaborn>=0.13

# Agent Framework
langgraph>=0.2
langchain>=0.3
langchain-openai>=0.2

# Configuration & Environment
python-dotenv>=1.0
pyyaml>=6.0

# Templating
jinja2>=3.1

# Testing
pytest>=8.0
